Fixed a bug in reward.php and got the index title to use the configured name
I'm not sold on keeping settings in the database like the current version does,
I'm considering moving all of those out to the config file. Either that or
several things currently in the config file get moved to the DB. It needs to
standardize one way or the other, and standardizing to the config file will
minimize DB queries per page.